Avoid bindable feedback causing overwrites
