+++ osu.Game.Rulesets.Taiko/Objects/TaikoHitObject.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Linq;
 using osu.Framework.Bindables;
+using osu.Game.Audio;
 using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Scoring;
@@ -35,15 +37,35 @@ namespace osu.Game.Rulesets.Taiko.Objects
         protected TaikoHitObject()
         {
             SamplesBindable.BindCollectionChanged((_, __) => UpdateTypeFromSamples());
-            TypeBindable.BindValueChanged(_ => UpdateSamplesFromType());
+            TypeBindable.BindValueChanged(_ => updateSamplesFromType());
         }
 
-        protected virtual void UpdateSamplesFromType()
+        private void updateSamplesFromType()
         {
+            var rimSamples = getRimSamples();
+
+            bool isRimType = Type == HitType.Rim;
+
+            if (isRimType != rimSamples.Any())
+            {
+                if (isRimType)
+                    Samples.Add(new HitSampleInfo(HitSampleInfo.HIT_CLAP));
+                else
+                {
+                    foreach (var sample in rimSamples)
+                        Samples.Remove(sample);
+                }
+            }
         }
 
         protected virtual void UpdateTypeFromSamples()
         {
+            Type = getRimSamples().Any() ? HitType.Rim : HitType.Centre;
         }
+
+        /// <summary>
+        /// Returns an array of any samples which would cause this object to be a "rim" type hit.
+        /// </summary>
+        private HitSampleInfo[] getRimSamples() => Samples.Where(s => s.Name == HitSampleInfo.HIT_CLAP || s.Name == HitSampleInfo.HIT_WHISTLE).ToArray();
     }
 }
